fix: update city interface defs to support versions
This adds a version key to each CityConfig option, and updates the getCityConfig() function to accept a version string that will be derived from the new route paths as a parameter (e.g. v1, v2)

This also updates the getCityConfig() function to return the version if the version key is found, otherwise throw an error that can be turned into a response for consistency in all handlers.