fix: refactor data to follow protocol/contract data
Third time's a charm! By separating out the variables by contract type (auth/core/token) it allows us to grab the relevant value to a contract (is auth activated?) as well as infer where to look for the data if there is a discrepancy.

This could also allow defining objects for the v1/v2/vN contracts that could be reused say, if the same auth/token contracts are used in a new version alongside a different core contract.